dashboard with task click works

+++ src/app/components/tache/tache.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit} from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output, OnChanges, SimpleChanges } from '@angular/core';
 import { Tache } from 'src/app/models/tache.model';
 import { TacheService } from 'src/app/service/tache.service';
 import { ActivatedRoute, Router } from '@angular/router';
@@ -10,10 +10,13 @@ type StatutTache = "ToDo" | "EnCours" | "Terminé" | "Test" | "Validé" | "Annul
   templateUrl: './tache.component.html',
   styleUrls: ['./tache.component.css']
 })
-export class TacheComponent implements OnInit {
-  tache: Tache | null = null;
+export class TacheComponent implements OnInit, OnChanges {
   editingDescription = false;
   tempDescription = '';
+  @Input() tache!: Tache;
+  @Output() fermer = new EventEmitter<void>();
+  @Output() delete = new EventEmitter<number>();
+  @Output() updateDescription = new EventEmitter<string>();
 
   constructor(
     private tacheService: TacheService,
@@ -23,8 +26,20 @@ export class TacheComponent implements OnInit {
 
   ngOnInit(): void {
     const id = this.route.snapshot.paramMap.get('id');
-    if (id) {
+    if (id && !this.tache) {
       this.chargerTache(+id);
+    }
+
+    if (this.tache) {
+      this.tempDescription = this.tache.descriptionTache;
+    }
+  }
+
+  ngOnChanges(changes: SimpleChanges): void {
+    // React to changes in the input tache
+    if (changes['tache'] && changes['tache'].currentValue) {
+      console.log('Tache changed:', this.tache);
+      this.tempDescription = this.tache.descriptionTache;
     }
   }
 
@@ -45,13 +60,19 @@ export class TacheComponent implements OnInit {
 
   editDescription(): void {
     this.editingDescription = true;
+    this.tempDescription = this.tache?.descriptionTache || '';
   }
 
   saveDescription(): void {
     if (this.tache) {
       this.editingDescription = false;
       this.tempDescription = this.tache.descriptionTache;
-      this.tacheService.updateTache(this.tache).subscribe();
+      this.updateDescription.emit(this.tache.descriptionTache);
+
+      // Only call service if we're in standalone mode (via router), not in popup mode
+      if (this.route.snapshot.paramMap.get('id')) {
+        this.tacheService.updateTache(this.tache).subscribe();
+      }
     }
   }
 
@@ -64,9 +85,15 @@ export class TacheComponent implements OnInit {
 
   deleteTache(): void {
     if (this.tache?.idTache) {
-      this.tacheService.deleteTache(this.tache.idTache).subscribe(() => {
-        this.router.navigate(['/']);
-      });
+      if (this.route.snapshot.paramMap.get('id')) {
+        // In standalone mode
+        this.tacheService.deleteTache(this.tache.idTache).subscribe(() => {
+          this.router.navigate(['/']);
+        });
+      } else {
+        // In popup mode
+        this.delete.emit(this.tache.idTache);
+      }
     }
   }
 
@@ -89,6 +116,12 @@ export class TacheComponent implements OnInit {
   }
 
   retour(): void {
-    this.router.navigate(['/']);
+    if (this.route.snapshot.paramMap.get('id')) {
+      // If we're on a standalone page
+      this.router.navigate(['/']);
+    } else {
+      // If we're in a popup
+      this.fermer.emit();
+    }
   }
 }
